moved Particpants folder to components folder
